+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Changed
+
+- ECOS: raise a ProblemError if the cost matrix is not positive definite
+
 ## [2.7.3] - 2023/01/16
 
 ### Added